@@ -6,7 +6,7 @@
 
 module k8s.io/autoscaler/cluster-autoscaler
 
-go 1.13
+go 1.15
 
 require (
 	cloud.google.com/go v0.51.0
