@@ -2,23 +2,22 @@
 
 module k8s.io/cloud-provider
 
-go 1.13
+go 1.15
 
 require (
 	github.com/google/go-cmp v0.4.0
 	github.com/stretchr/testify v1.4.0
-	k8s.io/api v0.0.0
-	k8s.io/apimachinery v0.0.0
-	k8s.io/client-go v0.0.0
-	k8s.io/component-base v0.0.0
+	k8s.io/api v0.19.16
+	k8s.io/apimachinery v0.19.16
+	k8s.io/client-go v0.19.16
+	k8s.io/component-base v0.19.16
 	k8s.io/klog/v2 v2.2.0
-	k8s.io/utils v0.0.0-20200619165400-6e3d28b6ed19
+	k8s.io/utils v0.0.0-20200729134348-d5654de09c73
 )
 
 replace (
-	k8s.io/api => ../api
-	k8s.io/apimachinery => ../apimachinery
-	k8s.io/client-go => ../client-go
-	k8s.io/cloud-provider => ../cloud-provider
-	k8s.io/component-base => ../component-base
+	k8s.io/api => k8s.io/api v0.19.16
+	k8s.io/apimachinery => k8s.io/apimachinery v0.19.16
+	k8s.io/client-go => k8s.io/client-go v0.19.16
+	k8s.io/component-base => k8s.io/component-base v0.19.16
 )
