@@ -2,21 +2,20 @@
 
 module k8s.io/csi-translation-lib
 
-go 1.13
+go 1.15
 
 require (
 	github.com/stretchr/testify v1.4.0
-	k8s.io/api v0.0.0
-	k8s.io/apimachinery v0.0.0
-	k8s.io/cloud-provider v0.0.0
+	k8s.io/api v0.19.16
+	k8s.io/apimachinery v0.19.16
+	k8s.io/cloud-provider v0.19.16
 	k8s.io/klog/v2 v2.2.0
 )
 
 replace (
-	k8s.io/api => ../api
-	k8s.io/apimachinery => ../apimachinery
-	k8s.io/client-go => ../client-go
-	k8s.io/cloud-provider => ../cloud-provider
-	k8s.io/component-base => ../component-base
-	k8s.io/csi-translation-lib => ../csi-translation-lib
+	k8s.io/api => k8s.io/api v0.19.16
+	k8s.io/apimachinery => k8s.io/apimachinery v0.19.16
+	k8s.io/client-go => k8s.io/client-go v0.19.16
+	k8s.io/cloud-provider => k8s.io/cloud-provider v0.19.16
+	k8s.io/component-base => k8s.io/component-base v0.19.16
 )
