@@ -52,8 +52,9 @@ type internalDeltaSnapshotData struct {
 	modifiedNodeInfoMap map[string]*schedulerframework.NodeInfo
 	deletedNodeInfos    map[string]bool
 
-	nodeInfoList         []*schedulerframework.NodeInfo
-	havePodsWithAffinity []*schedulerframework.NodeInfo
+	nodeInfoList                     []*schedulerframework.NodeInfo
+	havePodsWithAffinity             []*schedulerframework.NodeInfo
+	havePodsWithRequiredAntiAffinity []*schedulerframework.NodeInfo
 }
 
 func newInternalDeltaSnapshotData() *internalDeltaSnapshotData {
@@ -178,6 +179,7 @@ func (data *internalDeltaSnapshotData) clearCaches() {
 
 func (data *internalDeltaSnapshotData) clearPodCaches() {
 	data.havePodsWithAffinity = nil
+	data.havePodsWithRequiredAntiAffinity = nil
 }
 
 func (data *internalDeltaSnapshotData) removeNode(nodeName string) error {
@@ -324,6 +326,24 @@ func (snapshot *deltaSnapshotNodeLister) HavePodsWithAffinityList() ([]*schedule
 	}
 	data.havePodsWithAffinity = havePodsWithAffinityList
 	return data.havePodsWithAffinity, nil
+}
+
+// HavePodsWithRequiredAntiAffinityList returns list of all node infos with pods that have anti affinity constrints.
+func (snapshot *deltaSnapshotNodeLister) HavePodsWithRequiredAntiAffinityList() ([]*schedulerframework.NodeInfo, error) {
+	data := snapshot.data
+	if data.havePodsWithRequiredAntiAffinity != nil {
+		return data.havePodsWithRequiredAntiAffinity, nil
+	}
+
+	nodeInfoList := snapshot.data.getNodeInfoList()
+	havePodsWithRequiredAntiAffinityList := make([]*schedulerframework.NodeInfo, 0, len(nodeInfoList))
+	for _, node := range nodeInfoList {
+		if len(node.PodsWithRequiredAntiAffinity) > 0 {
+			havePodsWithRequiredAntiAffinityList = append(havePodsWithRequiredAntiAffinityList, node)
+		}
+	}
+	data.havePodsWithRequiredAntiAffinity = havePodsWithRequiredAntiAffinityList
+	return data.havePodsWithRequiredAntiAffinity, nil
 }
 
 // Get returns node info by node name.
